+++ MicroServices/Business/Business.Application.Contracts/BusinessApplicationContractsModule.cs
@@ -1,16 +1,34 @@
-﻿using Volo.Abp.Application;
+﻿using Business.Localization;
+using Volo.Abp.Application;
+using Volo.Abp.Localization;
 using Volo.Abp.Modularity;
 using Volo.Abp.PermissionManagement;
+using Volo.Abp.Validation.Localization;
+using Volo.Abp.VirtualFileSystem;
 
 namespace Business
 {
     [DependsOn(
-        typeof(BusinessDomainModule),
+        typeof(AbpLocalizationModule),
         typeof(AbpDddApplicationModule),
         typeof(AbpPermissionManagementApplicationContractsModule)
     )]
     public class BusinessApplicationContractsModule : AbpModule
     {
+        public override void ConfigureServices(ServiceConfigurationContext context)
+        {
+            Configure<AbpVirtualFileSystemOptions>(options =>
+            {
+                options.FileSets.AddEmbedded<BusinessApplicationContractsModule>("Business");
+            });
 
+            Configure<AbpLocalizationOptions>(options =>
+            {
+                options.Resources
+                    .Add<BusinessResource>("zh-Hans")
+                    .AddBaseTypes(typeof(AbpValidationResource))
+                    .AddVirtualJson("/Localization/Business");
+            });
+        }
     }
 }
