+++ MicroServices/Business/Business.Application.Contracts/Localization/BusinessResource.cs
