+++ MicroServices/Business/Business.Application.Contracts/Permissions/BusinessPermissionDefinitionProvider.cs
@@ -1,8 +1,4 @@
 ﻿using Business.Localization;
-using Microsoft.Extensions.Localization;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using Volo.Abp.Authorization.Permissions;
 using Volo.Abp.Localization;
 using Volo.Abp.MultiTenancy;
