+++ MicroServices/Business/Business.Core/Consts/BusinessConsts.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Business
+﻿namespace Business
 {
     public static class BusinessConsts
     {
