+++ MicroServices/Business/Business.Core/Consts/MultiTenancyConsts.cs
