+++ MicroServices/Business/Business.Domain/Business.Domain.csproj
@@ -19,20 +19,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Remove="Localization\Business\cs.json" />
-    <None Remove="Localization\Business\en.json" />
-    <None Remove="Localization\Business\zh-Hans.json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <EmbeddedResource Include="Localization\Business\cs.json" />
-    <EmbeddedResource Include="Localization\Business\en.json" />
-    <EmbeddedResource Include="Localization\Business\zh-Hans.json">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </EmbeddedResource>
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Volo.Abp.AuditLogging.Domain" Version="4.0.0" />
     <PackageReference Include="Volo.Abp.Ddd.Domain" Version="4.0.0" />
     <PackageReference Include="Volo.Abp.EntityFrameworkCore" Version="4.0.0" />
@@ -43,6 +29,10 @@
 
   <ItemGroup>
     <Folder Include="Models\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Business.Core\Business.Core.csproj" />
   </ItemGroup>
 
 </Project>
